remove Flyway completely from Gradle build (fixes FINERACT-765)
we (of course), still, but only, run it in-App at runtime boot

This is a clean-up which removes something we don't actually need.

Most importantly, it unblocks the Gradle upgrade in FINERACT-700,
where this doesn't work anymore anyway (and while there's a fix
making it work on one of my branches, it seems simpler to just
remove it).