rm git & JDK from Dockerfile
Now that we separated the build and runtime containers,
we can have only a JRE but not a JDK in the runtime container
(whereas the builder container will still have a JRE).

This makes for a smaller, and more importantly,
more secure, runtime container image.

Also removing git from runtime container, as it's not used.